First Commit
This first commit only add a general structure on one module (practice), and create a controller with an index method to serve the root request ('/'). It serve the root request with very simple reply 'Hello there'. It hide all the other code that used to be served from this path.